requirements.txt
psycopg2-binary>=2.9
python-dotenv>=1.0
torch>=2.2
numpy>=1.26
pandas>=2.2
flask>=3.0
gymnasium>=0.29
tqdm>=4.66
pytest>=8.0
